1.12.21

+++ src/main/java/com/mixplus/library/system/ProcessInfo.java
@@ -3,7 +3,7 @@ package com.mixplus.library.system;
 public record ProcessInfo(
         int pid,
         String name,
-        long memoryUsage,
+        double memoryUsage,
         int threadCount,
         String path
 ) { }
